feat(bash): exit code display and hidden-line count in TUI output
